Add tests for components
